requirements.txt
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
urllib3==2.3.0
python-dotenv==1.0.0
yop-cloud-sdk @ git+https://github.com/lqrhy3/yop-cloud-sdk.git@e27db97239a7f810bf70217b7f6445f16ea0db88
